% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glancer.R
\name{glancer}
\alias{glancer}
\title{Glancer}
\usage{
glancer(obj)
}
\arguments{
\item{obj}{A list of model fits}
}
\value{
A list of results from the broom.mixed:glance
}
\description{
Call glance on a list of model fits
}
