% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerModList_lt.R
\name{lmerModList_lt}
\alias{lmerModList_lt}
\title{lmerModList_lt}
\usage{
lmerModList_lt(formula, data, parallel = FALSE, ...)
}
\arguments{
\item{formula}{a formula to pass to lmer}

\item{data}{a list of data frames}

\item{parallel}{a logical variable indicating whether the models should be run
run in parallel}

\item{...}{pass additional arguments to lmer}
}
\value{
a list of fitted merMod objects (the class merModList has been removed)
}
\description{
Like lmerModList from the merTools package but calls lmerTest::lmer intead of
lme4::lmer
}
