% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFixedEff_x.R
\name{modelFixedEff_x}
\alias{modelFixedEff_x}
\title{modelFixedEff_x}
\usage{
modelFixedEff_x(modList, exponentiate = TRUE, ...)
}
\arguments{
\item{modList}{a merModList object}

\item{exponentiate}{if TRUE, coefficients will be exponentiated}

\item{...}{pass arguments to tidy}
}
\value{
output of merTools::modelFixedEff plus: median p-value
}
\description{
Adds to the output of merTools::modelFixedEff
}
\details{
The statistical literature suggests using the median p-value when conducting
logistic regression on multiple imputation data:

"In context of significance testing for logistic regression, Eekhout, Wiel,
and Heymans (2017) suggest taking the median of the mp-values as the combined
p-value, an exceedingly simple method. It nevertheless appears to outperform
more sophisticated techniques if the variable to be tested is categorical
with more than two categories. It would be useful to explore whether this
median P rule has wider validity."

https://stefvanbuuren.name/fimd/sec-multiparameter.html#sec:chi
}
